Update Master Filter
Added a global -is:masterwork, this will remove the need for keep tags on your pinnacles & ritual weapons although this does mean that non-well rolled MW'ed armor will not be filtered at all. I'd rather not have someone MW a piece of armor that they like and have it highlighted to be deleted.