Removing is:statlower
is:statlower currently cross compares blues and purples. I disagree with this method and am removing is:statlower from the Vault Analyzer. There is not enough known about blues for this method of comparison to be applied to every combination of blue and purple armor. 

Cross-comparing blues and purples has no positive functionality and can negatively impact new players / low-experience players.